Added javadoc commented files
Did not change anything in the files and I had wrote the javadoc
comments and generated the javadocs yesterday (you can see the javadocs
were already fully created) but I had forgotten to update the files in
this github repository with the javadoc comments. So here it is. I did
not do extra work since the deadline, only added the comments that were
clearly written before the deadline.